+++ CHANGELOG.md
@@ -841,18 +841,23 @@
 - reorganized tests
 - tests mostly or all sucessfull when ran individually
 - updates to remote webserver & tests
+**9/18/2023**
+- continued testing remote webserver
+
 
 ------------------------------------------------------------------------------------
 
 ## TODO
 
-- continue testing remote webserver
-- test remote webserver; add/update tests for remote webserver
+- check / test reconnect webdriver functionality
+- finish debugging remote webdriver
 
 - continue refactoring webdriver tests into relying only on webdriver components
 - and continue refactoring class tests into relying only on class components
 
 - RUN ALL TESTS AGAIN until successful
+
+- finish testing remote webserver; add/update tests for remote webserver
 
 
 - refactor specific class names into shared variables like: new_post_text_input --> message_or_post_textarea